add float64 cast for mne resample

+++ py_neuromodulation/nm_resample.py
@@ -47,4 +47,4 @@ class Resampler:
         """
         if not self.up:
             return data
-        return mne.filter.resample(data, up=self.up, down=1.0)
+        return mne.filter.resample(data.astype(np.float64), up=self.up, down=1.0)
